+++ src/server/endpoints/covidcast_nowcast.py
@@ -1,11 +1,13 @@
 from flask import Blueprint, request
 
 from .._query import execute_query, filter_integers, filter_strings
-from .._validate import (
+from .._params import (
     extract_date,
     extract_dates,
     extract_integer,
     extract_strings,
+)
+from .._validate import (
     require_all,
     require_any,
 )
