+++ src/server/endpoints/ght.py
@@ -2,7 +2,11 @@ from flask import Blueprint, request
 
 from .._config import AUTH
 from .._query import execute_query, QueryBuilder
-from .._validate import check_auth_token, extract_integers, extract_strings, require_all
+from .._params import (
+    extract_integers,
+    extract_strings,
+)
+from .._validate import check_auth_token, require_all
 
 # first argument is the endpoint name
 bp = Blueprint("ght", __name__)
