+++ src/server/endpoints/nidss_dengue.py
@@ -3,7 +3,11 @@ import re
 from flask import Blueprint
 
 from .._query import execute_queries, filter_integers
-from .._validate import extract_integers, extract_strings, require_all
+from .._params import (
+    extract_integers,
+    extract_strings,
+)
+from .._validate import require_all
 
 # first argument is the endpoint name
 bp = Blueprint("nidss_dengue", __name__)
