+++ src/server/endpoints/sensors.py
@@ -1,10 +1,12 @@
 from flask import Blueprint
 
 from .._config import AUTH, GRANULAR_SENSOR_AUTH_TOKENS, OPEN_SENSORS
+from .._params import (
+    extract_integers,
+    extract_strings,
+)
 from .._validate import (
     require_all,
-    extract_strings,
-    extract_integers,
     resolve_auth_token,
 )
 from .._query import filter_strings, execute_query, filter_integers
