+++ src/server/endpoints/twitter.py
@@ -2,10 +2,12 @@ from flask import Blueprint, request
 
 from .._config import AUTH, NATION_REGION, REGION_TO_STATE
 from .._query import execute_queries, filter_dates, filter_integers, filter_strings
-from .._validate import (
-    check_auth_token,
+from .._params import (
     extract_integers,
     extract_strings,
+)
+from .._validate import (
+    check_auth_token,
     require_all,
     require_any,
 )
